+++ OpenCL/dijkstra.cl
@@ -1,5 +1,5 @@
 // to initialize the buffer
-__kernel void initializeBuffers(__global int *maskArray, __global long long *costArray, __global long long *updatingCostArray, int sourceVertex, int vertexCount) {
+__kernel void initializeBuffers(__global int *maskArray, __global int *costArray, __global int *updatingCostArray, int sourceVertex, int vertexCount) {
     int tid = get_global_id(0);
 
     if(sourceVertex == tid) {
@@ -19,8 +19,8 @@ __kernel void DijkstraKernel1(__global int *vertexArray,
                               __global int *edgeArray,
                               __global int *weightArray,
                               __global int *maskArray,
-                              __global long long *costArray, 
-                              __global long long *updatingCostArray,
+                              __global int *costArray, 
+                              __global int *updatingCostArray,
                               int vertexCount, 
                               int edgeCount) {
     int tid = get_global_id(0);
@@ -48,8 +48,8 @@ __kernel void DijkstraKernel2(__global int *vertexArray,
                               __global int *edgeArray, 
                               __global int *weightArray, 
                               __global int *maskArray, 
-                              __global long long *costArray, 
-                              __global long long *updatingCostArray, 
+                              __global int *costArray, 
+                              __global int *updatingCostArray, 
                               int vertexCount) {
     int tid = get_global_id(0);
